IMPORTANT: the support for arbitrary big integers proves to expensive to use it for the general solving. It will become an option in a later version (with 2 executables built then).

+++ solver09zidsmt/aggs/AggTypes.h
@@ -14,7 +14,7 @@ using namespace tr1;
 
 #include "BigInteger.hh"
 #include "BigIntegerUtils.hh"
-typedef BigInteger Weight;
+typedef int Weight;
 
 namespace Aggrs{
 class WLV;
